refactor: App.jsx의 상태 관리 로직을 TodoContainer로 분리
